requirements.txt
environs~=9.5.0
python-telegram-bot~=13.13
vk-api~=11.9.9
